requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==4.0
django-cors-headers==3.11.0
django-neomodel==0.0.7
djangorestframework==3.13.1
interchange==2021.0.4
monotonic==1.6
neo4j==4.4.1
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
packaging==21.3
pansi==2020.7.3
py2neo==2021.2.3
Pygments==2.11.2
pyparsing==3.0.6
pytz==2021.3
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
